% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_utils.R
\name{afl_filter_from_expr}
\alias{afl_filter_from_expr}
\title{Convert filter expression(s) to AFL filter}
\usage{
afl_filter_from_expr(e)
}
\arguments{
\item{e}{An R expression vector of length 1 or more}
}
\value{
An AFL filter string
}
\description{
Convert filter expression(s) to AFL filter
}
