% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_utils.R
\name{afl_join_fields}
\alias{afl_join_fields}
\title{Just multple fields with sep = ','}
\usage{
afl_join_fields(..., sep = ",")
}
\arguments{
\item{...}{Multiple string vectors}
}
\description{
Default behavior: \code{paste(..., sep = sep, collapse = sep)} where \code{sep = ','}
}
\details{
afl(...) will convert vectors to joined strings separated by \verb{,}.
This function is useful in concatenating multiple vectors in parallel,
e.g. joining a new field vector and expression vector for the \code{apply} operator.
}
