% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_list.R
\docType{data}
\name{ArgList}
\alias{ArgList}
\title{ArgList}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
ArgList
}
\description{
Encapsulate arg list of function calls
}
\details{
Encapsulate arg list of function calls to avoid length argument passing among R functions
}
\keyword{datasets}
