% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_utils.R
\name{args_to_expressions}
\alias{args_to_expressions}
\title{Convert API ... args to an R expression vector}
\usage{
args_to_expressions(..., .param_list = NULL)
}
\arguments{
\item{...}{API ellipsis arg}

\item{.param_list}{Explicitly provide a parameter list. If not NULL, the ellipsis params are ignored}
}
\value{
R expression vector
}
\description{
Some API functions include ... arg to represent arbitrary search criteria.
This provides flexibility and simplifies API function signatures, but only supports limited advanced search,
e.g. xxx_contains, xxx_range, xxx_not.
}
\details{
Eg. name_contains = 'str' => name \%contains\% 'str'
Eg. value_range = c(1, 9) => c(value >= 1, value <= 9)
}
\seealso{
See \code{\link{e}} for more.
}
