% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{arrayop}
\alias{arrayop}
\title{ArrayOp: A package for Object-Oriented SciDB Array Operations/Operands}
\description{
ArrayOp package provides high-level, intutive API functions for data wrangling
and hide detailed AFL generation from you.
}
\details{
Synopsis
\enumerate{
\item Create to scidb
\item Get arrayOp isntances
\item Invoke verbs on arrayOp instances
\item Or in rare cases, manually create AFL statements
}
}
