% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_op_base.R
\docType{class}
\name{ArrayOpBase-class}
\alias{ArrayOpBase-class}
\alias{ArrayOpBase}
\title{Base class of all ArrayOp classes}
\arguments{
\item{field_names}{R character}

\item{selected_fields}{which fields are selected no matter what the parent operation is.
If NULL, .self fields will pass on by default depending on the parent operation.
Return AFL when .self used as an operand in another parent operation.}

\item{drop_dims}{Whether .self dimensions will be dropped in parent operations
By default, dimensions are not dropped in parent operation
But in some operations, dimensions are dropped or converted to attributes
e.g. equi_join creates two artificial dimensions and discard any existing dimensions of two operands.}
}
\value{
a named list as \code{field_names}, where absent fields or fields without data types are dropped silently.
Returns AFL when .self used as an operand in another parent operation without considering fields selection
}
\description{
Base class of all ArrayOp classes
}
\details{
NOTE: this method do not take into acount field selection
Sub-classes should override this function.
Doing so equips sub-classes with sensible behavior defined below.
Returns AFL when .self used as an operand in another parent operation.

By default, 1. dimensions are not dropped in parent operation; 2. no intent to select fields

implemented by calling .to_afl_explicit with \code{selected_fields = .self$get_field_names(.SEL)}
Return AFL suitable for retrieving data.frame.

scidb::iquery has a param \code{only_attributes}, which, if set TRUE, will effectively drop all dims.
Generate afl for ArrayOp used in a join context (equi_join).
}

