% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_op_v19.R
\name{ArrayOpV19}
\alias{ArrayOpV19}
\title{Base class of all ArrayOp classes}
\description{
ArrayOp classes denote scidb array operations and operands, hence the name.
}
\details{
One operation consists of an scidb operator and \link{1..*} operands, of which the result can be used as an operand
in another operation. Operands and Opreration results can all be denoted by ArrayOp.
}
\section{Super class}{
\code{\link[arrayop:ArrayOpBase]{arrayop::ArrayOpBase}} -> \code{ArrayOpV19}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{ArrayOpV19$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id=".get_meta">}\href{../../arrayop/html/ArrayOpBase.html#method-.get_meta}{\code{arrayop::ArrayOpBase$.get_meta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id=".set_meta">}\href{../../arrayop/html/ArrayOpBase.html#method-.set_meta}{\code{arrayop::ArrayOpBase$.set_meta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="array_meta_data">}\href{../../arrayop/html/ArrayOpBase.html#method-array_meta_data}{\code{arrayop::ArrayOpBase$array_meta_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="build_new">}\href{../../arrayop/html/ArrayOpBase.html#method-build_new}{\code{arrayop::ArrayOpBase$build_new()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="change_schema">}\href{../../arrayop/html/ArrayOpBase.html#method-change_schema}{\code{arrayop::ArrayOpBase$change_schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="create_new">}\href{../../arrayop/html/ArrayOpBase.html#method-create_new}{\code{arrayop::ArrayOpBase$create_new()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="create_new_with_same_schema">}\href{../../arrayop/html/ArrayOpBase.html#method-create_new_with_same_schema}{\code{arrayop::ArrayOpBase$create_new_with_same_schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="execute">}\href{../../arrayop/html/ArrayOpBase.html#method-execute}{\code{arrayop::ArrayOpBase$execute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="exists_persistent_array">}\href{../../arrayop/html/ArrayOpBase.html#method-exists_persistent_array}{\code{arrayop::ArrayOpBase$exists_persistent_array()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="filter">}\href{../../arrayop/html/ArrayOpBase.html#method-filter}{\code{arrayop::ArrayOpBase$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="finalize">}\href{../../arrayop/html/ArrayOpBase.html#method-finalize}{\code{arrayop::ArrayOpBase$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="initialize">}\href{../../arrayop/html/ArrayOpBase.html#method-initialize}{\code{arrayop::ArrayOpBase$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="inner_join">}\href{../../arrayop/html/ArrayOpBase.html#method-inner_join}{\code{arrayop::ArrayOpBase$inner_join()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="is_persistent">}\href{../../arrayop/html/ArrayOpBase.html#method-is_persistent}{\code{arrayop::ArrayOpBase$is_persistent()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="left_join">}\href{../../arrayop/html/ArrayOpBase.html#method-left_join}{\code{arrayop::ArrayOpBase$left_join()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="limit">}\href{../../arrayop/html/ArrayOpBase.html#method-limit}{\code{arrayop::ArrayOpBase$limit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="mutate">}\href{../../arrayop/html/ArrayOpBase.html#method-mutate}{\code{arrayop::ArrayOpBase$mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="mutate_by">}\href{../../arrayop/html/ArrayOpBase.html#method-mutate_by}{\code{arrayop::ArrayOpBase$mutate_by()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="mutate_old">}\href{../../arrayop/html/ArrayOpBase.html#method-mutate_old}{\code{arrayop::ArrayOpBase$mutate_old()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="overwrite">}\href{../../arrayop/html/ArrayOpBase.html#method-overwrite}{\code{arrayop::ArrayOpBase$overwrite()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="persist">}\href{../../arrayop/html/ArrayOpBase.html#method-persist}{\code{arrayop::ArrayOpBase$persist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="remove_self">}\href{../../arrayop/html/ArrayOpBase.html#method-remove_self}{\code{arrayop::ArrayOpBase$remove_self()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="remove_versions">}\href{../../arrayop/html/ArrayOpBase.html#method-remove_versions}{\code{arrayop::ArrayOpBase$remove_versions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="reshape">}\href{../../arrayop/html/ArrayOpBase.html#method-reshape}{\code{arrayop::ArrayOpBase$reshape()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="right_join">}\href{../../arrayop/html/ArrayOpBase.html#method-right_join}{\code{arrayop::ArrayOpBase$right_join()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="row_count">}\href{../../arrayop/html/ArrayOpBase.html#method-row_count}{\code{arrayop::ArrayOpBase$row_count()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="select">}\href{../../arrayop/html/ArrayOpBase.html#method-select}{\code{arrayop::ArrayOpBase$select()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="semi_join">}\href{../../arrayop/html/ArrayOpBase.html#method-semi_join}{\code{arrayop::ArrayOpBase$semi_join()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="set_auto_fields">}\href{../../arrayop/html/ArrayOpBase.html#method-set_auto_fields}{\code{arrayop::ArrayOpBase$set_auto_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="set_null_fields">}\href{../../arrayop/html/ArrayOpBase.html#method-set_null_fields}{\code{arrayop::ArrayOpBase$set_null_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="spawn">}\href{../../arrayop/html/ArrayOpBase.html#method-spawn}{\code{arrayop::ArrayOpBase$spawn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="summarize">}\href{../../arrayop/html/ArrayOpBase.html#method-summarize}{\code{arrayop::ArrayOpBase$summarize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="to_afl">}\href{../../arrayop/html/ArrayOpBase.html#method-to_afl}{\code{arrayop::ArrayOpBase$to_afl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="to_df">}\href{../../arrayop/html/ArrayOpBase.html#method-to_df}{\code{arrayop::ArrayOpBase$to_df()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="to_df_attrs">}\href{../../arrayop/html/ArrayOpBase.html#method-to_df_attrs}{\code{arrayop::ArrayOpBase$to_df_attrs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="to_schema_str">}\href{../../arrayop/html/ArrayOpBase.html#method-to_schema_str}{\code{arrayop::ArrayOpBase$to_schema_str()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="transmute">}\href{../../arrayop/html/ArrayOpBase.html#method-transmute}{\code{arrayop::ArrayOpBase$transmute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="update">}\href{../../arrayop/html/ArrayOpBase.html#method-update}{\code{arrayop::ArrayOpBase$update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="versions">}\href{../../arrayop/html/ArrayOpBase.html#method-versions}{\code{arrayop::ArrayOpBase$versions()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayOpV19$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
