% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Connect to SciDB server}
\usage{
connect(
  username,
  token,
  host = "127.0.0.1",
  port = 8083,
  protocol = "https",
  auth_type = "scidb",
  ...,
  save_to_default_conn = TRUE
)
}
\description{
Connect to SciDB server
}
