% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_op.R
\name{CustomizedOp}
\alias{CustomizedOp}
\title{CustomizedOp}
\description{
An ArrayOp sub-class with all properties customized for flexibility
}
\section{Super class}{
\code{\link[arrayop:ArrayOpBase]{arrayop::ArrayOpBase}} -> \code{CustomizedOp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomizedOp$new()}}
\item \href{#method-.raw_afl}{\code{CustomizedOp$.raw_afl()}}
\item \href{#method-get_absent_fields}{\code{CustomizedOp$get_absent_fields()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id=".to_afl_explicit">}\href{../../arrayop/html/ArrayOpBase.html#method-.to_afl_explicit}{\code{arrayop::ArrayOpBase$.to_afl_explicit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="convert_df">}\href{../../arrayop/html/ArrayOpBase.html#method-convert_df}{\code{arrayop::ArrayOpBase$convert_df()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="get_field_types">}\href{../../arrayop/html/ArrayOpBase.html#method-get_field_types}{\code{arrayop::ArrayOpBase$get_field_types()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="rename_fields">}\href{../../arrayop/html/ArrayOpBase.html#method-rename_fields}{\code{arrayop::ArrayOpBase$rename_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="to_afl">}\href{../../arrayop/html/ArrayOpBase.html#method-to_afl}{\code{arrayop::ArrayOpBase$to_afl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="to_df_afl">}\href{../../arrayop/html/ArrayOpBase.html#method-to_df_afl}{\code{arrayop::ArrayOpBase$to_df_afl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="to_join_operand_afl">}\href{../../arrayop/html/ArrayOpBase.html#method-to_join_operand_afl}{\code{arrayop::ArrayOpBase$to_join_operand_afl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="transform_unpack">}\href{../../arrayop/html/ArrayOpBase.html#method-transform_unpack}{\code{arrayop::ArrayOpBase$transform_unpack()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayop" data-topic="ArrayOpBase" data-id="validate_filter_expr">}\href{../../arrayop/html/ArrayOpBase.html#method-validate_filter_expr}{\code{arrayop::ArrayOpBase$validate_filter_expr()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomizedOp$new(
  afl_expr,
  dims = NULL,
  attrs = NULL,
  selected = NULL,
  field_types = list(),
  validate_fields = F,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.raw_afl"></a>}}
\subsection{Method \code{.raw_afl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomizedOp$.raw_afl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_absent_fields"></a>}}
\subsection{Method \code{get_absent_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomizedOp$get_absent_fields(fieldNames)}\if{html}{\out{</div>}}
}

}
}
