% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.R
\name{DbAccess}
\alias{DbAccess}
\title{An environment for scidb access}
\usage{
DbAccess(db)
}
\arguments{
\item{db}{scidb database connection object}
}
\description{
DbAccess shield the \code{arrayop} package from depending directly on \code{scidb}
}
\details{
Within \code{arrayop} package, all scidb transcactions are done via DbAccess.
No calls to \code{scidb::iquery} so that ArrayOp classes will not be affected by future scidb upgrade.
}
