% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_utils.R
\name{e}
\alias{e}
\title{Create a list of R expressions}
\usage{
e(...)
}
\arguments{
\item{...}{The ellipsis arg can have multiple items as expressions, but NO named items as in a named list.}
}
\value{
A list of R expressions
}
\description{
The ... ellipsis arg can include arbitrary expressions, where all names are preserved in their literal forms,
\strong{except} for those prefixed with !! (double exclamation marks) which will be evaluated to their actual values
in the calling environment.
}
\details{
Besides common comparison operators including \code{==}, \code{>}, \code{<}, \code{>=}, \code{<=}, \code{!=}, there are a few special operators
supported to ease AFL generation:
\itemize{
\item \code{\%in\%} semantically similar to R. \code{a \%in\% !!c(1,2,3)} will be translated to \code{(a == 1 or a == 2 or a == 3)}
\item \code{\%like\%} for string regex matching.
}
}
