% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_utils.R
\name{e_merge}
\alias{e_merge}
\title{Merge multiple R expressions into one}
\usage{
e_merge(el, mode = "AND")
}
\arguments{
\item{el}{A list of R expressions}

\item{mode}{'AND' | 'OR'. Logical relationships when merging the expressions.}
}
\value{
R expression
}
\description{
Merge an ExprsList into a single Expression so that it can be used as a FilterExpr
}
