% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_op.R
\name{get_aio_op}
\alias{get_aio_op}
\title{Get an CustomizedOp that generates AFL for aio_input}
\usage{
get_aio_op(
  filepath,
  template,
  aio_settings = list(),
  field_conversion = NULL,
  field_types
)
}
\arguments{
\item{filepath}{A single file path}

\item{template}{An ArrayOp sub-class used as an template for fields detection and matching}

\item{aio_settings}{Customized settings of aio_input}

\item{field_conversion}{If NULL (default), use template's field type to convert aio_input attributes; Or provide
a list for customized field conversion}

\item{field_types}{Instead of using a template, explicitly specify a field_types list. Cannot be used with template}
}
\value{
CustomizedOp ready to be used in a WriteOp as the dataset
}
\description{
Get an CustomizedOp that generates AFL for aio_input
}
