% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert.list}
\alias{invert.list}
\title{Invert names and values of a list (or named vector)}
\usage{
invert.list(obj, .as.list = T)
}
\arguments{
\item{obj}{a named list}

\item{.as.list}{Return a list if set TRUE; otherwise a named vector}
}
\value{
A list or named vector whose key/values are inverted from the original \code{obj}
}
\description{
Invert names and values of a list (or named vector)
}
