% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_op.R
\docType{class}
\name{MatchOp-class}
\alias{MatchOp-class}
\alias{MatchOp}
\title{MatchOp has the same schema as its operand. So all field queries are delegated to its operand.}
\description{
MatchOp has the same schema as its operand. So all field queries are delegated to its operand.
}

