% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_list.R
\name{newArgList}
\alias{newArgList}
\title{Create a ArgList isntance}
\usage{
newArgList(envir = parent.frame(), default_args = list())
}
\arguments{
\item{envir}{an environment, default as the calling environment}

\item{default_args}{A list of default arguments}
}
\value{
ArgList instance
}
\description{
Create a ArgList isntance from the calling environment
}
