% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_v19.R
\docType{data}
\name{RepoV19}
\alias{RepoV19}
\title{RepoBase class manages ArrayOp objects in one scidb installation}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
RepoV19
}
\arguments{
\item{namespace}{The default namespace}

\item{dbAccess}{A DbAccess instantance that manages scidb connection}
}
\description{
A repository for scidb arrays and derived ArrayOp objects that reside in one scidb installation

Initialize function.

Create a new Repo instance

Do NOT call this \code{initialize} function directly. Call \code{newRepo} instead to get a new Repo instance.
}
\details{
RepoBase is the base class of scidb-version-specific Repo classes.
Common ogic/behaviors across all supported scidb versions are placed here.
Version-specific logic/behaviors should be implemented in RepoBase sub-classes.
}
\keyword{datasets}
