% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_utils.R
\name{validate_filter_expr}
\alias{validate_filter_expr}
\title{Validate a filter expression ('filterExpr' must be a single R call expression)}
\usage{
validate_filter_expr(filterExpr, allFieldNames)
}
\arguments{
\item{filterExpr}{An rlang::expr}
}
\value{
A list object with named elements:
}
\description{
Current only report errors on:
\enumerate{
\item Name symbols that are not existing schema fields
\item Non-atomic 'values'
}
}
