% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_list.R
\name{ArgList}
\alias{ArgList}
\title{ArgList}
\description{
Encapsulate arg list of function calls
}
\details{
Encapsulate arg list of function calls to avoid length argument passing among R functions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ArgList$new()}}
\item \href{#method-get_len}{\code{ArgList$get_len()}}
\item \href{#method-has_args}{\code{ArgList$has_args()}}
\item \href{#method-has_all_args}{\code{ArgList$has_all_args()}}
\item \href{#method-has_any_args}{\code{ArgList$has_any_args()}}
\item \href{#method-has_args_count}{\code{ArgList$has_args_count()}}
\item \href{#method-has_scalar_args}{\code{ArgList$has_scalar_args()}}
\item \href{#method-has_multi_len_args}{\code{ArgList$has_multi_len_args()}}
\item \href{#method-args_count}{\code{ArgList$args_count()}}
\item \href{#method-subset}{\code{ArgList$subset()}}
\item \href{#method-rename}{\code{ArgList$rename()}}
\item \href{#method-assert_exclusive_args}{\code{ArgList$assert_exclusive_args()}}
\item \href{#method-to_exprs_list}{\code{ArgList$to_exprs_list()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$new(explicit_args, effective_args)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_len"></a>}}
\subsection{Method \code{get_len()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$get_len(explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_args"></a>}}
\subsection{Method \code{has_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$has_args(argNames, explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_all_args"></a>}}
\subsection{Method \code{has_all_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$has_all_args(argNames, explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_any_args"></a>}}
\subsection{Method \code{has_any_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$has_any_args(argNames, explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_args_count"></a>}}
\subsection{Method \code{has_args_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$has_args_count(argNames, count, explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_scalar_args"></a>}}
\subsection{Method \code{has_scalar_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$has_scalar_args(argNames, explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_multi_len_args"></a>}}
\subsection{Method \code{has_multi_len_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$has_multi_len_args(argNames, explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-args_count"></a>}}
\subsection{Method \code{args_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$args_count(argNames, count, explicit = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$subset(argNames, inverse = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rename"></a>}}
\subsection{Method \code{rename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$rename(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_exclusive_args"></a>}}
\subsection{Method \code{assert_exclusive_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$assert_exclusive_args(
  argNames,
  errorTemplate = "Argument '\%s' cannot be specified with '\%s'"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_exprs_list"></a>}}
\subsection{Method \code{to_exprs_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArgList$to_exprs_list(explicit = FALSE)}\if{html}{\out{</div>}}
}

}
}
