% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{arrayop}
\alias{arrayop}
\title{ArrayOp: A package for Object-Oriented SciDB Array Operations/Operands}
\description{
ArrayOp abstracts away AFL generation by unifying array operations and array operands.
}
\section{ArrayOpBase class and its sub-classes}{

content1
}

\section{Repo classes}{

content2
}

