% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_op_v19.R
\docType{data}
\name{ArrayOpV19}
\alias{ArrayOpV19}
\title{Base class of all ArrayOp classes}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
ArrayOpV19
}
\description{
ArrayOp classes denote scidb array operations and operands, hence the name.
}
\details{
One operation consists of an scidb operator and \link{1..*} operands, of which the result can be used as an operand
in another operation. Operands and Opreration results can all be denoted by ArrayOp.
}
\keyword{datasets}
