% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DBUtils}
\alias{DBUtils}
\title{Database utility class}
\description{
Database utility class

Database utility class
}
\details{
Common scidb operations
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-random_array_name}{\code{DBUtils$random_array_name()}}
\item \href{#method-random_field_name}{\code{DBUtils$random_field_name()}}
\item \href{#method-set_conn}{\code{DBUtils$set_conn()}}
\item \href{#method-get_conn}{\code{DBUtils$get_conn()}}
\item \href{#method-clear_cache}{\code{DBUtils$clear_cache()}}
\item \href{#method-list_arrays_in_ns}{\code{DBUtils$list_arrays_in_ns()}}
\item \href{#method-load_array_ops_from_namespace}{\code{DBUtils$load_array_ops_from_namespace()}}
\item \href{#method-sanitize_names}{\code{DBUtils$sanitize_names()}}
\item \href{#method-sanitize_names_for}{\code{DBUtils$sanitize_names_for()}}
\item \href{#method-db_namespaces}{\code{DBUtils$db_namespaces()}}
\item \href{#method-db_users}{\code{DBUtils$db_users()}}
\item \href{#method-db_roles}{\code{DBUtils$db_roles()}}
\item \href{#method-db_operators}{\code{DBUtils$db_operators()}}
\item \href{#method-db_instances}{\code{DBUtils$db_instances()}}
\item \href{#method-db_queries}{\code{DBUtils$db_queries()}}
\item \href{#method-db_macros}{\code{DBUtils$db_macros()}}
\item \href{#method-db_types}{\code{DBUtils$db_types()}}
\item \href{#method-db_libraries}{\code{DBUtils$db_libraries()}}
\item \href{#method-db_aggregates}{\code{DBUtils$db_aggregates()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_array_name"></a>}}
\if{latex}{\out{\hypertarget{method-random_array_name}{}}}
\subsection{Method \code{random_array_name()}}{
Generate a random array name so it does not collides with existing array names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$random_array_name(prefix = "Rarrayop_", n = 10L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_field_name"></a>}}
\if{latex}{\out{\hypertarget{method-random_field_name}{}}}
\subsection{Method \code{random_field_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$random_field_name(n = 10L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_conn"></a>}}
\if{latex}{\out{\hypertarget{method-set_conn}{}}}
\subsection{Method \code{set_conn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$set_conn(new_conn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_conn"></a>}}
\if{latex}{\out{\hypertarget{method-get_conn}{}}}
\subsection{Method \code{get_conn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$get_conn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_arrays_in_ns"></a>}}
\if{latex}{\out{\hypertarget{method-list_arrays_in_ns}{}}}
\subsection{Method \code{list_arrays_in_ns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$list_arrays_in_ns(ns = "public")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_array_ops_from_namespace"></a>}}
\if{latex}{\out{\hypertarget{method-load_array_ops_from_namespace}{}}}
\subsection{Method \code{load_array_ops_from_namespace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$load_array_ops_from_namespace(ns = "public")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sanitize_names"></a>}}
\if{latex}{\out{\hypertarget{method-sanitize_names}{}}}
\subsection{Method \code{sanitize_names()}}{
Sanitize (data frame or scidb array) names

First replace any non-alphanumerical letter to _
Then trim off any leading or trailing underscores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$sanitize_names(original_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{original_names}}{A character vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of sanitized names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sanitize_names_for"></a>}}
\if{latex}{\out{\hypertarget{method-sanitize_names_for}{}}}
\subsection{Method \code{sanitize_names_for()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$sanitize_names_for(obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-db_namespaces}{}}}
\subsection{Method \code{db_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_users"></a>}}
\if{latex}{\out{\hypertarget{method-db_users}{}}}
\subsection{Method \code{db_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_users()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_roles"></a>}}
\if{latex}{\out{\hypertarget{method-db_roles}{}}}
\subsection{Method \code{db_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_roles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_operators"></a>}}
\if{latex}{\out{\hypertarget{method-db_operators}{}}}
\subsection{Method \code{db_operators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_operators()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_instances"></a>}}
\if{latex}{\out{\hypertarget{method-db_instances}{}}}
\subsection{Method \code{db_instances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_instances()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_queries"></a>}}
\if{latex}{\out{\hypertarget{method-db_queries}{}}}
\subsection{Method \code{db_queries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_queries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_macros"></a>}}
\if{latex}{\out{\hypertarget{method-db_macros}{}}}
\subsection{Method \code{db_macros()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_macros()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_types"></a>}}
\if{latex}{\out{\hypertarget{method-db_types}{}}}
\subsection{Method \code{db_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_libraries"></a>}}
\if{latex}{\out{\hypertarget{method-db_libraries}{}}}
\subsection{Method \code{db_libraries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_libraries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_aggregates"></a>}}
\if{latex}{\out{\hypertarget{method-db_aggregates}{}}}
\subsection{Method \code{db_aggregates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$db_aggregates()}\if{html}{\out{</div>}}
}

}
}
