% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DBUtils}
\alias{DBUtils}
\title{Database utility class}
\description{
Database utility class

Database utility class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-random_array_name}{\code{DBUtils$random_array_name()}}
\item \href{#method-random_field_name}{\code{DBUtils$random_field_name()}}
\item \href{#method-set_conn}{\code{DBUtils$set_conn()}}
\item \href{#method-get_conn}{\code{DBUtils$get_conn()}}
\item \href{#method-clear_cache}{\code{DBUtils$clear_cache()}}
\item \href{#method-list_arrays_in_ns}{\code{DBUtils$list_arrays_in_ns()}}
\item \href{#method-load_array_ops_from_namespace}{\code{DBUtils$load_array_ops_from_namespace()}}
\item \href{#method-sanitize_names}{\code{DBUtils$sanitize_names()}}
\item \href{#method-sanitize_names_for}{\code{DBUtils$sanitize_names_for()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_array_name"></a>}}
\if{latex}{\out{\hypertarget{method-random_array_name}{}}}
\subsection{Method \code{random_array_name()}}{
Generate a random array name so it does not collides with existing array names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$random_array_name(prefix = "Rarrayop_", n = 10L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_field_name"></a>}}
\if{latex}{\out{\hypertarget{method-random_field_name}{}}}
\subsection{Method \code{random_field_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$random_field_name(n = 10L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_conn"></a>}}
\if{latex}{\out{\hypertarget{method-set_conn}{}}}
\subsection{Method \code{set_conn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$set_conn(new_conn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_conn"></a>}}
\if{latex}{\out{\hypertarget{method-get_conn}{}}}
\subsection{Method \code{get_conn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$get_conn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_arrays_in_ns"></a>}}
\if{latex}{\out{\hypertarget{method-list_arrays_in_ns}{}}}
\subsection{Method \code{list_arrays_in_ns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$list_arrays_in_ns(ns = "public")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_array_ops_from_namespace"></a>}}
\if{latex}{\out{\hypertarget{method-load_array_ops_from_namespace}{}}}
\subsection{Method \code{load_array_ops_from_namespace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$load_array_ops_from_namespace(ns = "public")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sanitize_names"></a>}}
\if{latex}{\out{\hypertarget{method-sanitize_names}{}}}
\subsection{Method \code{sanitize_names()}}{
Sanitize (data frame or scidb array) names

First replace any non-alphanumerical letter to _
Then trim off any leading or trailing underscores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$sanitize_names(original_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{original_names}}{A character vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of sanitized names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sanitize_names_for"></a>}}
\if{latex}{\out{\hypertarget{method-sanitize_names_for}{}}}
\subsection{Method \code{sanitize_names_for()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBUtils$sanitize_names_for(obj)}\if{html}{\out{</div>}}
}

}
}
