% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_default_connection}
\alias{get_default_connection}
\title{Get the default ScidbConnection}
\usage{
get_default_connection(.report_error_if_not_connected = TRUE)
}
\description{
Call \code{arrayop::db_connect} to establish a connection
}
