% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_base.R
\name{newRepo}
\alias{newRepo}
\title{Create a scidb-version-specific Repo instance}
\usage{
newRepo(default_namespace = "ns", dependency_obj = NULL, db)
}
\arguments{
\item{dependency_obj}{A named list that contains 4 functions: get_scidb_version, query, execute and get_schema_df
If left NULL, will require a scidb connection 'db' to create the dependency_obj automatically.}

\item{db}{a scidb connection returned by scidb::scidbconnect}
}
\description{
The created Repo instance is automatically selected based on Repo-Scidb compatibility
}
\details{
This should be only way to create a new Repo instance,
unless a specific Repo version is known/desired beforehand, which is useful in test cases.
}
