% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{Repo}
\alias{Repo}
\title{Repo class which manages ArraySchemas that reside in one scidb installation}
\description{
A repository for ArraySchemas that reside in one scidb installation
}
\details{
Repo class depends on a package-level dependency object 'DEP'
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Repo$new()}}
\item \href{#method-to_df}{\code{Repo$to_df()}}
\item \href{#method-run_only}{\code{Repo$run_only()}}
\item \href{#method-register_schema}{\code{Repo$register_schema()}}
\item \href{#method-load_schema_from_cache}{\code{Repo$load_schema_from_cache()}}
\item \href{#method-load_schema_from_db}{\code{Repo$load_schema_from_db()}}
\item \href{#method-remove_array_versions}{\code{Repo$remove_array_versions()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize function

Create a new Repo instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repo$new(namespace, dbAccess = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{The default namespace}

\item{\code{dbAccess}}{A DbAccess instantance that manages scidb connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_df"></a>}}
\subsection{Method \code{to_df()}}{
Get a data frame from any ArrayOp instance or raw AFL

\strong{NOTE} if \code{only_attributes = T}, then \code{op$to_afl(drop_dims = T)} which may differ from the default \code{op$to_afl()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repo$to_df(what, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{An ArrayOp instance or raw AFL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_only"></a>}}
\subsection{Method \code{run_only()}}{
Run scidb statements without expecting results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repo$run_only(what, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{An ArrayOp instance or raw AFL
If \code{what} is an \code{ArrayOp} instance, \code{what$to_afl()} will be used as the raw AFL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_schema"></a>}}
\subsection{Method \code{register_schema()}}{
Register an array schema into the Repo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repo$register_schema(alias, array_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{a user friendly name for a scidb array}

\item{\code{array_name}}{the actual array name in scidb}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_schema_from_cache"></a>}}
\subsection{Method \code{load_schema_from_cache()}}{
Load schema from Repo's cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repo$load_schema_from_cache(schema_alias)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{schema_alias}}{Alias of the array, not actual array name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cached ArraySchema or NULL if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_schema_from_db"></a>}}
\subsection{Method \code{load_schema_from_db()}}{
Load an array schema from scidb.
NOTE: No caching involved, always load from scidb
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repo$load_schema_from_db(array_name, ns = private$namespace)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array_name}}{Array name without namespace. Not array alias.}

\item{\code{namespace}}{Scidb namespace, which defaults to the Repo's namespace}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_array_versions"></a>}}
\subsection{Method \code{remove_array_versions()}}{
Remove versions of array snapshots
If only \code{alias} is specified, remove all but the most recent version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repo$remove_array_versions(alias, version_id, keep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{a user friendly name for a scidb array}

\item{\code{version_id}}{remove all versions up the \code{version_id}}

\item{\code{keep}}{remove all but the last \code{keep} versions}
}
\if{html}{\out{</div>}}
}
}
}
