% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_base.R
\name{RepoBase}
\alias{RepoBase}
\title{RepoBase class manages ArrayOp objects in one scidb installation}
\description{
A repository for scidb arrays and derived ArrayOp objects that reside in one scidb installation
}
\details{
RepoBase is the base class of scidb-version-specific Repo classes.
Common ogic/behaviors across all supported scidb versions are placed here.
Version-specific logic/behaviors should be implemented in RepoBase sub-classes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RepoBase$new()}}
\item \href{#method-query}{\code{RepoBase$query()}}
\item \href{#method-execute}{\code{RepoBase$execute()}}
\item \href{#method-ArrayOp}{\code{RepoBase$ArrayOp()}}
\item \href{#method-register_schema_alias_by_array_name}{\code{RepoBase$register_schema_alias_by_array_name()}}
\item \href{#method-get_alias_schema}{\code{RepoBase$get_alias_schema()}}
\item \href{#method-load_schema_from_db}{\code{RepoBase$load_schema_from_db()}}
\item \href{#method-clone}{\code{RepoBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize function.

Create a new Repo instance

Do NOT call this \code{initialize} function directly. Call \code{newRepo} instead to get a new Repo instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$new(default_namespace, dependency_object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{The default namespace}

\item{\code{dbAccess}}{A DbAccess instantance that manages scidb connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\subsection{Method \code{query()}}{
Run a query and expect results

Delegate to Repo's \code{dep$query} function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$query(afl, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{afl}}{Scidb AFL statement}

\item{\code{...}}{Extra settings passed to \code{dep$query} function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
Run a command and do not expect results

Delegate to Repo's \code{dep$execute} function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$execute(afl, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{afl}}{Scidb AFL statement}

\item{\code{...}}{Extra settings passed to \code{dep$execute} function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayOp"></a>}}
\subsection{Method \code{ArrayOp()}}{
Create a new ArrayOp instance with the proper scidb version

Should be overriden by sub-classes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$ArrayOp(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_schema_alias_by_array_name"></a>}}
\subsection{Method \code{register_schema_alias_by_array_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$register_schema_alias_by_array_name(
  alias,
  array_name,
  is_full_name = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_alias_schema"></a>}}
\subsection{Method \code{get_alias_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$get_alias_schema(alias)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_schema_from_db"></a>}}
\subsection{Method \code{load_schema_from_db()}}{
Load Array Schema directly from scidb
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$load_schema_from_db(full_array_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
