% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_base.R
\docType{data}
\name{RepoBase}
\alias{RepoBase}
\title{RepoBase class manages ArrayOp objects in one scidb installation}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
RepoBase
}
\arguments{
\item{namespace}{The default namespace}

\item{dbAccess}{A DbAccess instantance that manages scidb connection}

\item{what}{An ArrayOp instance or raw AFL}

\item{...}{Extra settings passed to \code{dep$query} function}

\item{what}{An ArrayOp instance or raw AFL}

\item{...}{Extra settings passed to \code{dep$execute} function}
}
\description{
A repository for scidb arrays and derived ArrayOp objects that reside in one scidb installation

Initialize function.

Create a new Repo instance

Do NOT call this \code{initialize} function directly. Call \code{newRepo} instead to get a new Repo instance.

Run a query and expect results

Delegate to Repo's \code{dep$query} function
Get a data frame from any ArrayOp instance or raw AFL

\strong{NOTE} if \code{only_attributes = T}, then \code{op$to_afl(drop_dims = T)} which may differ from the default \code{op$to_afl()}

Run a command and do not expect results

Delegate to Repo's \code{dep$execute} function

Create a new ArrayOp instance with the proper scidb version

Should be overriden by sub-classes

Register a scidb array as an alias in the Repo

If array_name is not fully qualified, the current default_namespace meta info is taken.
Later changes to the default_namespace will not affect already registered arrays

Load a new ArrayOp directly from scidb

The repo's \code{dep$get_schema_df} function is called to get array schema denoted by a data frame

Set Repo meta data directly

Get Repo meta data directly
}
\details{
RepoBase is the base class of scidb-version-specific Repo classes.
Common ogic/behaviors across all supported scidb versions are placed here.
Version-specific logic/behaviors should be implemented in RepoBase sub-classes.
}
\keyword{datasets}
