% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_v19.R
\name{RepoV19}
\alias{RepoV19}
\title{RepoBase class manages ArrayOp objects in one scidb installation}
\description{
A repository for scidb arrays and derived ArrayOp objects that reside in one scidb installation
}
\details{
RepoBase is the base class of scidb-version-specific Repo classes.
Common ogic/behaviors across all supported scidb versions are placed here.
Version-specific logic/behaviors should be implemented in RepoBase sub-classes.
}
\section{Super class}{
\code{\link[arrayopTemp:RepoBase]{arrayopTemp::RepoBase}} -> \code{RepoV19}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RepoV19$new()}}
\item \href{#method-ArrayOp}{\code{RepoV19$ArrayOp()}}
\item \href{#method-clone}{\code{RepoV19$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="arrayopTemp" data-topic="RepoBase" data-id="execute">}\href{../../arrayopTemp/html/RepoBase.html#method-execute}{\code{arrayopTemp::RepoBase$execute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayopTemp" data-topic="RepoBase" data-id="get_alias_schema">}\href{../../arrayopTemp/html/RepoBase.html#method-get_alias_schema}{\code{arrayopTemp::RepoBase$get_alias_schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayopTemp" data-topic="RepoBase" data-id="load_schema_from_db">}\href{../../arrayopTemp/html/RepoBase.html#method-load_schema_from_db}{\code{arrayopTemp::RepoBase$load_schema_from_db()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayopTemp" data-topic="RepoBase" data-id="query">}\href{../../arrayopTemp/html/RepoBase.html#method-query}{\code{arrayopTemp::RepoBase$query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="arrayopTemp" data-topic="RepoBase" data-id="register_schema_alias_by_array_name">}\href{../../arrayopTemp/html/RepoBase.html#method-register_schema_alias_by_array_name}{\code{arrayopTemp::RepoBase$register_schema_alias_by_array_name()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize function.

Create a new Repo instance

Do NOT call this \code{initialize} function directly. Call \code{newRepo} instead to get a new Repo instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoV19$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{The default namespace}

\item{\code{dbAccess}}{A DbAccess instantance that manages scidb connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayOp"></a>}}
\subsection{Method \code{ArrayOp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoV19$ArrayOp(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepoV19$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
