% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_db.R
\name{run_tests_with_scidb_connection}
\alias{run_tests_with_scidb_connection}
\title{Run tests with a scidb connection}
\usage{
run_tests_with_scidb_connection(db, namespace)
}
\arguments{
\item{db}{a scidb connection returned by calling \code{scidb::scidbconnect}}

\item{namespace}{Namespace to run tests in}
}
\description{
Run tests with a scidb connection
}
