% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{utility}
\alias{utility}
\title{An R env that hosts arrayop package utility functions}
\format{
An object of class \code{environment} of length 11.
}
\usage{
utility
}
\description{
An R env that hosts arrayop package utility functions
}
\keyword{datasets}
